refactor(udp): extract linux module
(cherry picked from commit 35ba3c45112bafa8b2f3963444f57fa58b10bcee)
